Extract a test thread factory
This is different from the production thread factory:
 - No daemon threads in tests. We don't need 'em.
 - Threads are numbered in tests. We do need that.
